# shared-contracts/target-operational-v2.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReportSubmissionCommand": {
        "additionalProperties": false,
        "properties": {
          "commandId": {
            "format": "uuid",
            "title": "Commandid",
            "type": "string"
          },
          "contractVersion": {
            "const": 2,
            "title": "Contractversion",
            "type": "integer"
          },
          "expectedVersion": {
            "minimum": 0.0,
            "title": "Expectedversion",
            "type": "integer"
          },
          "idempotencyKey": {
            "maxLength": 240,
            "minLength": 8,
            "pattern": "^report:[A-Za-z0-9._-]+:[A-Za-z0-9._-]+$",
            "title": "Idempotencykey",
            "type": "string"
          },
          "occurredAt": {
            "format": "date-time",
            "title": "Occurredat",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/ReportSubmissionCommandPayload"
          }
        },
        "required": [
          "contractVersion",
          "commandId",
          "idempotencyKey",
          "occurredAt",
          "expectedVersion",
          "payload"
        ],
        "title": "ReportSubmissionCommand",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "TANAW API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/operational/desktop/report-submissions/v2": {
      "post": {
        "operationId": "ingest_report_submission_v2_operational_desktop_report_submissions_v2_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportSubmissionCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportSubmissionAcknowledgement"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ingest Report Submission V2",
        "tags": [
          "reporting-v2"
        ]
      }
    },
...
    "/operational/enterprise/reports/v2": {
      "get": {
        "operationId": "list_owned_enterprise_reports_v2_operational_enterprise_reports_v2_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1024,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "reportingPeriodId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reportingperiodid"
            }
          },
          {
            "in": "query",
            "name": "workflowState",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "submitted",
                    "returned",
                    "accepted",
                    "consolidated"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workflowstate"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseReportPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Owned Enterprise Reports V2",
        "tags": [
          "reporting-v2"
        ]
      }
    },
    "/operational/enterprise/reports/{enterprise_report_id}/v2": {
      "get": {
        "operationId": "read_owned_enterprise_report_v2_operational_enterprise_reports__enterprise_report_id__v2_get",
        "parameters": [
          {
            "in": "path",
            "name": "enterprise_report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Enterprise Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseReportDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Owned Enterprise Report V2",
        "tags": [
          "reporting-v2"
        ]
      }
    },
...
    "/operational/reporting-periods/lifecycle/run/v2": {
      "post": {
        "operationId": "run_reporting_period_lifecycle_v2_operational_reporting_periods_lifecycle_run_v2_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportingPeriodLifecycleResult"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Run Reporting Period Lifecycle V2",
        "tags": [
          "reporting-v2"
        ]
      }
    },
    "/operational/reporting-periods/v2": {
      "get": {
        "operationId": "list_reporting_periods_v2_operational_reporting_periods_v2_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1024,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "periodStatus",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "scheduled",
                    "open",
                    "closed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Periodstatus"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportingPeriodPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Reporting Periods V2",
        "tags": [
          "reporting-v2"
        ]
      }
    },
    "/operational/reporting-periods/{reporting_period_id}/compliance/v2": {
      "get": {
        "operationId": "read_reporting_period_compliance_v2_operational_reporting_periods__reporting_period_id__compliance_v2_get",
        "parameters": [
          {
            "in": "path",
            "name": "reporting_period_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reporting Period Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeriodComplianceResource"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Reporting Period Compliance V2",
        "tags": [
          "reporting-v2"
        ]
      }
    },
    "/operational/reporting-periods/{reporting_period_id}/obligations/freeze/v2": {
      "post": {
        "operationId": "freeze_reporting_period_obligations_v2_operational_reporting_periods__reporting_period_id__obligations_freeze_v2_post",
        "parameters": [
          {
            "in": "path",
            "name": "reporting_period_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reporting Period Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObligationFreezeCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObligationFreezeAcknowledgement"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Freeze Reporting Period Obligations V2",
        "tags": [
          "reporting-v2"
        ]
      }
    },
    "/operational/reporting-periods/{reporting_period_id}/reminder-intents/v2": {
      "post": {
        "operationId": "create_reporting_period_reminder_intents_v2_operational_reporting_periods__reporting_period_id__reminder_intents_v2_post",
        "parameters": [
          {
            "in": "path",
            "name": "reporting_period_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reporting Period Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderIntentCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderIntentAcknowledgement"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Reporting Period Reminder Intents V2",
        "tags": [
          "reporting-v2"
        ]
      }
    },
    "/operational/reporting-periods/{reporting_period_id}/v2": {
      "get": {
        "operationId": "read_reporting_period_v2_operational_reporting_periods__reporting_period_id__v2_get",
        "parameters": [
          {
            "in": "path",
            "name": "reporting_period_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reporting Period Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportingPeriodDiscoveryResource"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Reporting Period V2",
        "tags": [
          "reporting-v2"
        ]
      }
    },
...
    "/operational/reports/{enterprise_report_id}/transitions/v2": {
      "post": {
        "operationId": "transition_enterprise_report_v2_operational_reports__enterprise_report_id__transitions_v2_post",
        "parameters": [
          {
            "in": "path",
            "name": "enterprise_report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Enterprise Report Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportTransitionCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportTransitionAcknowledgement"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Transition Enterprise Report V2",
        "tags": [
          "reporting-v2"
        ]
      }
    },
    "/operational/reports/{enterprise_report_id}/v2": {
      "get": {
        "operationId": "read_enterprise_report_v2_operational_reports__enterprise_report_id__v2_get",
        "parameters": [
          {
            "in": "path",
            "name": "enterprise_report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Enterprise Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseReportDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Enterprise Report V2",
        "tags": [
          "reporting-v2"
        ]
      }
    },
...
  }
}